Revert "ui changes"
