+++ next.config.ts
@@ -2,7 +2,7 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
     experimental: {
-            turbo: {
+        turbo: {
             rules: {
                 "*.svg": {
                     loaders: ["@svgr/webpack"],
@@ -26,14 +26,7 @@ const nextConfig: NextConfig = {
         '@gravity-ui/uikit',
         // Remove @gravity-ui/uikit from here
     ],
-    reactStrictMode: false,
-    swcMinify: true,
-    eslint: {
-        ignoreDuringBuilds: true
-    },
-    typescript: {
-        ignoreBuildErrors: true
-    }
+    reactStrictMode: true,
 };
 
 export default nextConfig;
