+++ src/components/logoLoad/LogoLoader.tsx
@@ -1,52 +1,53 @@
 export const LogoLoad = () => {
     return (
-        <svg
-            id="logo"
-            width="300"
-            height="48"
-            viewBox="0 0 1104 174"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-        >
-            <g clipPath="url(#clip0_1_2)">
-                <path d="M20.12 104.429H0.679993V99.6288L36.44 24.0288H58.04L93.8 99.6288V104.429H74.36L67.4 88.8288H26.96L20.12 104.429ZM35 70.8288H59.48L47.24 43.3488L35 70.8288Z" />
-                <path d="M168.8 24.0288H190.4V70.9488C190.4 75.9088 189.76 80.5488 188.48 84.8688C187.2 89.1888 185 92.9888 181.88 96.2688C178.76 99.5488 174.56 102.129 169.28 104.009C164 105.889 157.36 106.829 149.36 106.829H145.04C137.04 106.829 130.4 105.889 125.12 104.009C119.84 102.129 115.64 99.5488 112.52 96.2688C109.4 92.9888 107.2 89.1888 105.92 84.8688C104.64 80.5488 104 75.9088 104 70.9488V24.0288H125.6V67.8288C125.6 70.4688 125.72 72.9688 125.96 75.3288C126.2 77.6888 126.98 79.7888 128.3 81.6288C129.62 83.4688 131.74 84.9288 134.66 86.0088C137.58 87.0888 141.76 87.6288 147.2 87.6288C152.56 87.6288 156.72 87.0888 159.68 86.0088C162.64 84.9288 164.78 83.4688 166.1 81.6288C167.42 79.7888 168.2 77.6888 168.44 75.3288C168.68 72.9688 168.8 70.4688 168.8 67.8288V24.0288Z" />
-                <path d="M289.4 104.429H268.52L252.44 79.2288C248.12 79.2288 243.96 79.1888 239.96 79.1088C235.96 79.0288 232.04 78.6688 228.2 78.0288V104.429H206.6V24.0288C213.48 23.2288 220.7 22.6288 228.26 22.2288C235.82 21.8288 243.24 21.6288 250.52 21.6288H253.88C260.6 21.6288 266.16 22.3288 270.56 23.7288C274.96 25.1288 278.48 27.0288 281.12 29.4288C283.76 31.8288 285.6 34.6288 286.64 37.8288C287.68 41.0288 288.2 44.4288 288.2 48.0288V52.8288C288.2 57.8688 287.2 62.4488 285.2 66.5688C283.2 70.6888 279.64 73.8688 274.52 76.1088L289.4 99.6288V104.429ZM250.76 40.8288C247.96 40.8288 244.5 40.9088 240.38 41.0688C236.26 41.2288 232.2 41.5088 228.2 41.9088V58.9488C231.96 59.4288 235.52 59.7288 238.88 59.8488C242.24 59.9688 244.96 60.0288 247.04 60.0288C251.84 60.0288 255.56 59.8088 258.2 59.3688C260.84 58.9288 262.78 58.2688 264.02 57.3888C265.26 56.5088 266 55.3888 266.24 54.0288C266.48 52.6688 266.6 51.0288 266.6 49.1088C266.6 48.0688 266.52 47.0488 266.36 46.0488C266.2 45.0488 265.62 44.1488 264.62 43.3488C263.62 42.5488 262.04 41.9288 259.88 41.4888C257.72 41.0488 254.68 40.8288 250.76 40.8288Z" />
-                <path d="M368.48 24.0288H390.08V70.9488C390.08 75.9088 389.44 80.5488 388.16 84.8688C386.88 89.1888 384.68 92.9888 381.56 96.2688C378.44 99.5488 374.24 102.129 368.96 104.009C363.68 105.889 357.04 106.829 349.04 106.829H344.72C336.72 106.829 330.08 105.889 324.8 104.009C319.52 102.129 315.32 99.5488 312.2 96.2688C309.08 92.9888 306.88 89.1888 305.6 84.8688C304.32 80.5488 303.68 75.9088 303.68 70.9488V24.0288H325.28V67.8288C325.28 70.4688 325.4 72.9688 325.64 75.3288C325.88 77.6888 326.66 79.7888 327.98 81.6288C329.3 83.4688 331.42 84.9288 334.34 86.0088C337.26 87.0888 341.44 87.6288 346.88 87.6288C352.24 87.6288 356.4 87.0888 359.36 86.0088C362.32 84.9288 364.46 83.4688 365.78 81.6288C367.1 79.7888 367.88 77.6888 368.12 75.3288C368.36 72.9688 368.48 70.4688 368.48 67.8288V24.0288Z" />
-                <path d="M461.48 87.9888H450.44L427.88 56.4288V104.429H406.28V24.0288H427.88L455.96 65.1888L484.04 24.0288H505.64V104.429H484.04V56.4288L461.48 87.9888Z" />
-                <path d="M560.96 87.6288C565.28 87.6288 568.66 87.4088 571.1 86.9688C573.54 86.5288 575.38 85.9688 576.62 85.2888C577.86 84.6088 578.64 83.8488 578.96 83.0088C579.28 82.1688 579.44 81.3888 579.44 80.6688C579.44 79.2288 579.1 78.0888 578.42 77.2488C577.74 76.4088 576.66 75.7488 575.18 75.2688C573.7 74.7888 571.74 74.4488 569.3 74.2488C566.86 74.0488 563.92 73.9088 560.48 73.8288C558.48 73.7488 556.12 73.6288 553.4 73.4688C550.68 73.3088 547.88 72.9688 545 72.4488C542.12 71.9288 539.28 71.1688 536.48 70.1688C533.68 69.1688 531.18 67.7688 528.98 65.9688C526.78 64.1688 525 61.9088 523.64 59.1888C522.28 56.4688 521.6 53.1488 521.6 49.2288V47.6688C521.6 44.2288 522.22 40.9288 523.46 37.7688C524.7 34.6088 526.82 31.8288 529.82 29.4288C532.82 27.0288 536.86 25.1288 541.94 23.7288C547.02 22.3288 553.36 21.6288 560.96 21.6288H565.28C570.8 21.6288 576.16 21.9688 581.36 22.6488C586.56 23.3288 591.64 24.3088 596.6 25.5888V45.3888C590.76 44.0288 584.98 42.9488 579.26 42.1488C573.54 41.3488 567.72 40.9488 561.8 40.9488C557.48 40.9488 554.08 41.1688 551.6 41.6088C549.12 42.0488 547.26 42.6088 546.02 43.2888C544.78 43.9688 544 44.6888 543.68 45.4488C543.36 46.2088 543.2 46.9488 543.2 47.6688C543.2 49.1088 543.54 50.2688 544.22 51.1488C544.9 52.0288 546 52.7088 547.52 53.1888C549.04 53.6688 551 54.0088 553.4 54.2088C555.8 54.4088 558.68 54.5488 562.04 54.6288C563.96 54.7088 566.24 54.8088 568.88 54.9288C571.52 55.0488 574.3 55.3688 577.22 55.8888C580.14 56.4088 583.02 57.1888 585.86 58.2288C588.7 59.2688 591.24 60.6888 593.48 62.4888C595.72 64.2888 597.54 66.5688 598.94 69.3288C600.34 72.0888 601.04 75.4288 601.04 79.3488V80.9088C601.04 84.3488 600.42 87.6288 599.18 90.7488C597.94 93.8688 595.8 96.6288 592.76 99.0288C589.72 101.429 585.68 103.329 580.64 104.729C575.6 106.129 569.24 106.829 561.56 106.829H557.24C550.92 106.829 544.78 106.489 538.82 105.809C532.86 105.129 527 104.189 521.24 102.989V82.8288C527.72 84.2688 534.42 85.4288 541.34 86.3088C548.26 87.1888 554.8 87.6288 560.96 87.6288Z" />
-                <path d="M616.64 24.0288H638.24V55.2288H652.04L678.2 24.0288H699.32V28.8288L669.68 64.2288L699.32 99.6288V104.429H678.2L653 74.4288H638.24V104.429H616.64V24.0288Z" />
-                <path d="M741.2 104.429V74.5488L709.76 28.8288V24.0288H731L752 57.3888L773.12 24.0288H794.24V28.8288L762.8 74.5488V104.429H741.2Z" />
-                <path d="M826.64 24.0288L868.76 75.6288V24.0288H890.36V104.429H868.76L826.64 52.8288V104.429H805.04V24.0288H826.64Z" />
-                <path d="M984.08 24.0288V43.2288H928.76V55.2288H972.08V74.4288H928.76V85.2288H984.08V104.429H907.16V24.0288H984.08Z" />
-                <path d="M1048.4 104.429H1026.8V43.2288H995.6V24.0288H1079.72V43.2288H1048.4V104.429Z" />
-                <path d="M214.65 172.402V141.662C214.65 139.875 214.9 138.288 215.4 136.902C215.907 135.522 216.7 134.358 217.78 133.412C218.867 132.458 220.287 131.728 222.04 131.222C223.793 130.722 225.94 130.472 228.48 130.472C229.52 130.472 230.527 130.518 231.5 130.612C232.467 130.705 233.323 130.845 234.07 131.032V139.202C233.323 138.975 232.493 138.815 231.58 138.722C230.667 138.628 229.78 138.582 228.92 138.582C227.653 138.582 226.667 138.675 225.96 138.862C225.247 139.048 224.713 139.328 224.36 139.702C224.007 140.075 223.793 140.515 223.72 141.022C223.647 141.522 223.61 142.108 223.61 142.782V144.412H232.23V152.242H223.61V172.402H214.65Z" />
-                <path d="M261.5 143.682V152.132C259.787 151.832 258.053 151.682 256.3 151.682C253.273 151.682 250.53 152.075 248.07 152.862V172.402H239.11V144.412H241.35L244.37 146.202C246.163 145.228 248.067 144.498 250.08 144.012C252.1 143.532 254.34 143.292 256.8 143.292C257.66 143.292 258.48 143.328 259.26 143.402C260.047 143.475 260.793 143.568 261.5 143.682Z" />
-                <path d="M267.1 159.522V157.282C267.1 155.342 267.343 153.522 267.83 151.822C268.317 150.128 269.173 148.645 270.4 147.372C271.633 146.105 273.277 145.108 275.33 144.382C277.383 143.655 279.977 143.292 283.11 143.292H284.68C287.813 143.292 290.407 143.655 292.46 144.382C294.513 145.108 296.157 146.105 297.39 147.372C298.617 148.645 299.473 150.128 299.96 151.822C300.447 153.522 300.69 155.342 300.69 157.282V159.522C300.69 161.462 300.447 163.282 299.96 164.982C299.473 166.682 298.617 168.165 297.39 169.432C296.157 170.698 294.513 171.695 292.46 172.422C290.407 173.155 287.813 173.522 284.68 173.522H283.11C279.977 173.522 277.383 173.155 275.33 172.422C273.277 171.695 271.633 170.698 270.4 169.432C269.173 168.165 268.317 166.682 267.83 164.982C267.343 163.282 267.1 161.462 267.1 159.522ZM276.06 158.402C276.06 159.295 276.097 160.172 276.17 161.032C276.243 161.892 276.533 162.638 277.04 163.272C277.54 163.905 278.313 164.418 279.36 164.812C280.407 165.205 281.92 165.402 283.9 165.402C285.84 165.402 287.34 165.205 288.4 164.812C289.467 164.418 290.25 163.905 290.75 163.272C291.257 162.638 291.547 161.892 291.62 161.032C291.693 160.172 291.73 159.295 291.73 158.402C291.73 157.508 291.693 156.632 291.62 155.772C291.547 154.912 291.257 154.165 290.75 153.532C290.25 152.898 289.467 152.385 288.4 151.992C287.34 151.598 285.84 151.402 283.9 151.402C281.92 151.402 280.407 151.598 279.36 151.992C278.313 152.385 277.54 152.898 277.04 153.532C276.533 154.165 276.243 154.912 276.17 155.772C276.097 156.632 276.06 157.508 276.06 158.402Z" />
-                <path d="M307.69 144.412H309.93L312.67 146.032C314.497 145.138 316.437 144.458 318.49 143.992C320.543 143.525 322.763 143.292 325.15 143.292C327.73 143.292 329.903 143.572 331.67 144.132C333.443 144.692 334.927 145.455 336.12 146.422C337.987 145.375 339.947 144.592 342 144.072C344.053 143.552 346.313 143.292 348.78 143.292C351.727 143.292 354.18 143.655 356.14 144.382C358.1 145.108 359.647 146.095 360.78 147.342C361.92 148.595 362.723 150.078 363.19 151.792C363.657 153.512 363.89 155.342 363.89 157.282V172.402H354.93V158.402C354.93 157.508 354.893 156.642 354.82 155.802C354.747 154.962 354.513 154.215 354.12 153.562C353.727 152.908 353.103 152.385 352.25 151.992C351.39 151.598 350.177 151.402 348.61 151.402C346.97 151.402 345.43 151.525 343.99 151.772C342.55 152.012 341.16 152.375 339.82 152.862C340.007 153.568 340.127 154.295 340.18 155.042C340.24 155.788 340.27 156.535 340.27 157.282V172.402H331.31V158.402C331.31 157.508 331.273 156.642 331.2 155.802C331.127 154.962 330.893 154.215 330.5 153.562C330.107 152.908 329.48 152.385 328.62 151.992C327.767 151.598 326.553 151.402 324.98 151.402C321.92 151.402 319.14 151.852 316.64 152.752V172.402H307.69V144.412Z" />
-                <path d="M400.11 165.962C401.53 165.962 402.65 165.905 403.47 165.792C404.29 165.678 404.907 165.538 405.32 165.372C405.727 165.205 405.987 165.018 406.1 164.812C406.213 164.605 406.27 164.392 406.27 164.172C406.27 163.312 405.757 162.788 404.73 162.602C403.703 162.415 402.07 162.285 399.83 162.212C399.083 162.212 398.197 162.175 397.17 162.102C396.143 162.022 395.08 161.888 393.98 161.702C392.88 161.515 391.79 161.235 390.71 160.862C389.623 160.488 388.66 159.968 387.82 159.302C386.98 158.628 386.3 157.788 385.78 156.782C385.26 155.775 385 154.542 385 153.082V152.522C385 151.295 385.233 150.128 385.7 149.022C386.16 147.922 386.97 146.942 388.13 146.082C389.29 145.228 390.83 144.548 392.75 144.042C394.67 143.542 397.087 143.292 400 143.292H401.57C403.697 143.292 405.767 143.412 407.78 143.652C409.793 143.892 411.753 144.238 413.66 144.692V152.472C411.42 151.945 409.2 151.542 407 151.262C404.793 150.982 402.517 150.842 400.17 150.842C398.71 150.842 397.57 150.898 396.75 151.012C395.93 151.125 395.313 151.275 394.9 151.462C394.493 151.648 394.233 151.845 394.12 152.052C394.007 152.252 393.95 152.465 393.95 152.692C393.95 153.478 394.453 153.972 395.46 154.172C396.473 154.378 398.117 154.518 400.39 154.592C401.097 154.632 401.963 154.672 402.99 154.712C404.017 154.745 405.08 154.865 406.18 155.072C407.287 155.278 408.38 155.568 409.46 155.942C410.54 156.315 411.51 156.835 412.37 157.502C413.23 158.175 413.92 159.015 414.44 160.022C414.96 161.035 415.22 162.268 415.22 163.722V164.282C415.22 165.515 414.987 166.682 414.52 167.782C414.06 168.882 413.25 169.862 412.09 170.722C410.93 171.575 409.39 172.255 407.47 172.762C405.55 173.268 403.133 173.522 400.22 173.522H398.65C396.23 173.522 393.88 173.398 391.6 173.152C389.327 172.912 387.087 172.568 384.88 172.122V164.332C387.347 164.858 389.893 165.262 392.52 165.542C395.153 165.822 397.683 165.962 400.11 165.962Z" />
-                <path d="M438.74 173.522H437.67C434.537 173.522 431.943 173.155 429.89 172.422C427.837 171.695 426.197 170.698 424.97 169.432C423.737 168.165 422.877 166.682 422.39 164.982C421.903 163.282 421.66 161.462 421.66 159.522V157.282C421.66 155.342 421.903 153.522 422.39 151.822C422.877 150.128 423.737 148.645 424.97 147.372C426.197 146.105 427.837 145.108 429.89 144.382C431.943 143.655 434.537 143.292 437.67 143.292H439.24C442.413 143.292 445.007 143.525 447.02 143.992C449.033 144.458 450.6 145.102 451.72 145.922C452.84 146.742 453.607 147.692 454.02 148.772C454.427 149.858 454.63 151.015 454.63 152.242V152.802C454.63 153.922 454.49 154.985 454.21 155.992C453.93 156.998 453.27 157.895 452.23 158.682C451.183 159.468 449.643 160.085 447.61 160.532C445.577 160.978 442.807 161.202 439.3 161.202C437.62 161.202 436.07 161.128 434.65 160.982C433.23 160.828 431.887 160.565 430.62 160.192C430.66 160.978 430.79 161.698 431.01 162.352C431.237 163.005 431.657 163.555 432.27 164.002C432.883 164.448 433.75 164.792 434.87 165.032C435.99 165.278 437.483 165.402 439.35 165.402C440.25 165.402 441.25 165.365 442.35 165.292C443.45 165.212 444.587 165.108 445.76 164.982C446.94 164.848 448.097 164.678 449.23 164.472C450.37 164.272 451.427 164.022 452.4 163.722V172.122C451.427 172.342 450.37 172.535 449.23 172.702C448.097 172.875 446.923 173.025 445.71 173.152C444.497 173.285 443.293 173.378 442.1 173.432C440.9 173.492 439.78 173.522 438.74 173.522ZM439.3 150.282C437.807 150.282 436.557 150.368 435.55 150.542C434.537 150.708 433.713 150.958 433.08 151.292C432.447 151.625 431.963 152.035 431.63 152.522C431.29 153.008 431.047 153.568 430.9 154.202C431.98 154.502 433.147 154.718 434.4 154.852C435.647 154.978 437.093 155.042 438.74 155.042C440.38 155.042 441.667 154.978 442.6 154.852C443.533 154.718 444.223 154.548 444.67 154.342C445.117 154.135 445.397 153.912 445.51 153.672C445.623 153.432 445.68 153.178 445.68 152.912C445.68 152.652 445.633 152.365 445.54 152.052C445.447 151.732 445.183 151.442 444.75 151.182C444.323 150.922 443.68 150.708 442.82 150.542C441.967 150.368 440.793 150.282 439.3 150.282Z" />
-                <path d="M462.19 130.412H464.43L471.15 134.332V172.402H462.19V130.412Z" />
-                <path d="M478.99 130.412H481.23L487.94 134.332V172.402H478.99V130.412Z" />
-                <path d="M512.29 173.522H511.23C508.097 173.522 505.503 173.155 503.45 172.422C501.397 171.695 499.753 170.698 498.52 169.432C497.293 168.165 496.437 166.682 495.95 164.982C495.463 163.282 495.22 161.462 495.22 159.522V157.282C495.22 155.342 495.463 153.522 495.95 151.822C496.437 150.128 497.293 148.645 498.52 147.372C499.753 146.105 501.397 145.108 503.45 144.382C505.503 143.655 508.097 143.292 511.23 143.292H512.8C515.973 143.292 518.567 143.525 520.58 143.992C522.593 144.458 524.16 145.102 525.28 145.922C526.4 146.742 527.167 147.692 527.58 148.772C527.987 149.858 528.19 151.015 528.19 152.242V152.802C528.19 153.922 528.05 154.985 527.77 155.992C527.49 156.998 526.83 157.895 525.79 158.682C524.743 159.468 523.203 160.085 521.17 160.532C519.137 160.978 516.363 161.202 512.85 161.202C511.17 161.202 509.623 161.128 508.21 160.982C506.79 160.828 505.447 160.565 504.18 160.192C504.213 160.978 504.343 161.698 504.57 162.352C504.79 163.005 505.21 163.555 505.83 164.002C506.443 164.448 507.31 164.792 508.43 165.032C509.55 165.278 511.043 165.402 512.91 165.402C513.803 165.402 514.8 165.365 515.9 165.292C517.007 165.212 518.147 165.108 519.32 164.982C520.493 164.848 521.65 164.678 522.79 164.472C523.93 164.272 524.983 164.022 525.95 163.722V172.122C524.983 172.342 523.93 172.535 522.79 172.702C521.65 172.875 520.473 173.025 519.26 173.152C518.047 173.285 516.843 173.378 515.65 173.432C514.457 173.492 513.337 173.522 512.29 173.522ZM512.85 150.282C511.357 150.282 510.107 150.368 509.1 150.542C508.093 150.708 507.273 150.958 506.64 151.292C506.007 151.625 505.52 152.035 505.18 152.522C504.847 153.008 504.607 153.568 504.46 154.202C505.54 154.502 506.707 154.718 507.96 154.852C509.207 154.978 510.65 155.042 512.29 155.042C513.937 155.042 515.227 154.978 516.16 154.852C517.093 154.718 517.783 154.548 518.23 154.342C518.677 154.135 518.957 153.912 519.07 153.672C519.183 153.432 519.24 153.178 519.24 152.912C519.24 152.652 519.193 152.365 519.1 152.052C519.007 151.732 518.743 151.442 518.31 151.182C517.883 150.922 517.24 150.708 516.38 150.542C515.52 150.368 514.343 150.282 512.85 150.282Z" />
-                <path d="M557.58 143.682V152.132C555.867 151.832 554.133 151.682 552.38 151.682C549.353 151.682 546.61 152.075 544.15 152.862V172.402H535.19V144.412H537.43L540.45 146.202C542.243 145.228 544.147 144.498 546.16 144.012C548.173 143.532 550.413 143.292 552.88 143.292C553.74 143.292 554.56 143.328 555.34 143.402C556.127 143.475 556.873 143.568 557.58 143.682Z" />
-                <path d="M586.13 137.132V144.412H594.75V152.242H586.13V161.202C586.13 161.835 586.167 162.395 586.24 162.882C586.32 163.368 586.527 163.778 586.86 164.112C587.193 164.445 587.717 164.698 588.43 164.872C589.137 165.038 590.123 165.122 591.39 165.122C592.25 165.122 593.147 165.075 594.08 164.982C595.013 164.888 595.853 164.728 596.6 164.502V172.682C595.853 172.868 594.993 173.008 594.02 173.102C593.053 173.195 592.047 173.242 591 173.242C588.467 173.242 586.32 172.998 584.56 172.512C582.807 172.025 581.39 171.325 580.31 170.412C579.23 169.498 578.437 168.358 577.93 166.992C577.423 165.632 577.17 164.075 577.17 162.322V133.212H579.41L586.13 137.132Z" />
-                <path d="M602.7 159.522V157.282C602.7 155.342 602.943 153.522 603.43 151.822C603.917 150.128 604.773 148.645 606 147.372C607.233 146.105 608.877 145.108 610.93 144.382C612.983 143.655 615.577 143.292 618.71 143.292H620.28C623.413 143.292 626.007 143.655 628.06 144.382C630.113 145.108 631.757 146.105 632.99 147.372C634.217 148.645 635.073 150.128 635.56 151.822C636.047 153.522 636.29 155.342 636.29 157.282V159.522C636.29 161.462 636.047 163.282 635.56 164.982C635.073 166.682 634.217 168.165 632.99 169.432C631.757 170.698 630.113 171.695 628.06 172.422C626.007 173.155 623.413 173.522 620.28 173.522H618.71C615.577 173.522 612.983 173.155 610.93 172.422C608.877 171.695 607.233 170.698 606 169.432C604.773 168.165 603.917 166.682 603.43 164.982C602.943 163.282 602.7 161.462 602.7 159.522ZM611.66 158.402C611.66 159.295 611.697 160.172 611.77 161.032C611.843 161.892 612.133 162.638 612.64 163.272C613.14 163.905 613.913 164.418 614.96 164.812C616.007 165.205 617.52 165.402 619.5 165.402C621.44 165.402 622.94 165.205 624 164.812C625.067 164.418 625.85 163.905 626.35 163.272C626.857 162.638 627.147 161.892 627.22 161.032C627.293 160.172 627.33 159.295 627.33 158.402C627.33 157.508 627.293 156.632 627.22 155.772C627.147 154.912 626.857 154.165 626.35 153.532C625.85 152.898 625.067 152.385 624 151.992C622.94 151.598 621.44 151.402 619.5 151.402C617.52 151.402 616.007 151.598 614.96 151.992C613.913 152.385 613.14 152.898 612.64 153.532C612.133 154.165 611.843 154.912 611.77 155.772C611.697 156.632 611.66 157.508 611.66 158.402Z" />
-                <path d="M672.51 165.962C673.93 165.962 675.05 165.905 675.87 165.792C676.69 165.678 677.303 165.538 677.71 165.372C678.123 165.205 678.387 165.018 678.5 164.812C678.613 164.605 678.67 164.392 678.67 164.172C678.67 163.312 678.157 162.788 677.13 162.602C676.103 162.415 674.47 162.285 672.23 162.212C671.483 162.212 670.597 162.175 669.57 162.102C668.543 162.022 667.48 161.888 666.38 161.702C665.28 161.515 664.187 161.235 663.1 160.862C662.02 160.488 661.06 159.968 660.22 159.302C659.38 158.628 658.7 157.788 658.18 156.782C657.653 155.775 657.39 154.542 657.39 153.082V152.522C657.39 151.295 657.623 150.128 658.09 149.022C658.557 147.922 659.37 146.942 660.53 146.082C661.683 145.228 663.223 144.548 665.15 144.042C667.07 143.542 669.487 143.292 672.4 143.292H673.96C676.093 143.292 678.167 143.412 680.18 143.652C682.193 143.892 684.153 144.238 686.06 144.692V152.472C683.82 151.945 681.597 151.542 679.39 151.262C677.19 150.982 674.913 150.842 672.56 150.842C671.107 150.842 669.97 150.898 669.15 151.012C668.33 151.125 667.713 151.275 667.3 151.462C666.893 151.648 666.633 151.845 666.52 152.052C666.407 152.252 666.35 152.465 666.35 152.692C666.35 153.478 666.853 153.972 667.86 154.172C668.867 154.378 670.51 154.518 672.79 154.592C673.497 154.632 674.363 154.672 675.39 154.712C676.417 154.745 677.48 154.865 678.58 155.072C679.68 155.278 680.773 155.568 681.86 155.942C682.94 156.315 683.91 156.835 684.77 157.502C685.63 158.175 686.32 159.015 686.84 160.022C687.36 161.035 687.62 162.268 687.62 163.722V164.282C687.62 165.515 687.387 166.682 686.92 167.782C686.453 168.882 685.643 169.862 684.49 170.722C683.33 171.575 681.79 172.255 679.87 172.762C677.95 173.268 675.533 173.522 672.62 173.522H671.05C668.623 173.522 666.273 173.398 664 173.152C661.72 172.912 659.48 172.568 657.28 172.122V164.332C659.747 164.858 662.293 165.262 664.92 165.542C667.553 165.822 670.083 165.962 672.51 165.962Z" />
-                <path d="M711.13 173.522H710.07C706.937 173.522 704.343 173.155 702.29 172.422C700.237 171.695 698.593 170.698 697.36 169.432C696.133 168.165 695.277 166.682 694.79 164.982C694.303 163.282 694.06 161.462 694.06 159.522V157.282C694.06 155.342 694.303 153.522 694.79 151.822C695.277 150.128 696.133 148.645 697.36 147.372C698.593 146.105 700.237 145.108 702.29 144.382C704.343 143.655 706.937 143.292 710.07 143.292H711.64C714.813 143.292 717.407 143.525 719.42 143.992C721.433 144.458 723 145.102 724.12 145.922C725.24 146.742 726.007 147.692 726.42 148.772C726.827 149.858 727.03 151.015 727.03 152.242V152.802C727.03 153.922 726.89 154.985 726.61 155.992C726.33 156.998 725.67 157.895 724.63 158.682C723.583 159.468 722.043 160.085 720.01 160.532C717.977 160.978 715.203 161.202 711.69 161.202C710.017 161.202 708.47 161.128 707.05 160.982C705.63 160.828 704.287 160.565 703.02 160.192C703.053 160.978 703.183 161.698 703.41 162.352C703.637 163.005 704.057 163.555 704.67 164.002C705.283 164.448 706.15 164.792 707.27 165.032C708.39 165.278 709.883 165.402 711.75 165.402C712.643 165.402 713.643 165.365 714.75 165.292C715.85 165.212 716.987 165.108 718.16 164.982C719.333 164.848 720.49 164.678 721.63 164.472C722.77 164.272 723.823 164.022 724.79 163.722V172.122C723.823 172.342 722.77 172.535 721.63 172.702C720.49 172.875 719.313 173.025 718.1 173.152C716.893 173.285 715.69 173.378 714.49 173.432C713.297 173.492 712.177 173.522 711.13 173.522ZM711.69 150.282C710.203 150.282 708.953 150.368 707.94 150.542C706.933 150.708 706.113 150.958 705.48 151.292C704.847 151.625 704.363 152.035 704.03 152.522C703.69 153.008 703.447 153.568 703.3 154.202C704.38 154.502 705.547 154.718 706.8 154.852C708.047 154.978 709.49 155.042 711.13 155.042C712.777 155.042 714.067 154.978 715 154.852C715.933 154.718 716.623 154.548 717.07 154.342C717.517 154.135 717.797 153.912 717.91 153.672C718.023 153.432 718.08 153.178 718.08 152.912C718.08 152.652 718.033 152.365 717.94 152.052C717.847 151.732 717.583 151.442 717.15 151.182C716.723 150.922 716.08 150.708 715.22 150.542C714.36 150.368 713.183 150.282 711.69 150.282Z" />
-                <path d="M734.59 130.412H736.83L743.55 134.332V172.402H734.59V130.412Z" />
-                <path d="M751.38 130.412H753.62L760.34 134.332V172.402H751.38V130.412Z" />
-                <path d="M784.69 173.522H783.63C780.497 173.522 777.903 173.155 775.85 172.422C773.797 171.695 772.153 170.698 770.92 169.432C769.687 168.165 768.83 166.682 768.35 164.982C767.863 163.282 767.62 161.462 767.62 159.522V157.282C767.62 155.342 767.863 153.522 768.35 151.822C768.83 150.128 769.687 148.645 770.92 147.372C772.153 146.105 773.797 145.108 775.85 144.382C777.903 143.655 780.497 143.292 783.63 143.292H785.2C788.367 143.292 790.96 143.525 792.98 143.992C794.993 144.458 796.56 145.102 797.68 145.922C798.8 146.742 799.567 147.692 799.98 148.772C800.387 149.858 800.59 151.015 800.59 152.242V152.802C800.59 153.922 800.45 154.985 800.17 155.992C799.89 156.998 799.227 157.895 798.18 158.682C797.14 159.468 795.603 160.085 793.57 160.532C791.53 160.978 788.757 161.202 785.25 161.202C783.57 161.202 782.023 161.128 780.61 160.982C779.19 160.828 777.847 160.565 776.58 160.192C776.613 160.978 776.743 161.698 776.97 162.352C777.19 163.005 777.61 163.555 778.23 164.002C778.843 164.448 779.71 164.792 780.83 165.032C781.95 165.278 783.443 165.402 785.31 165.402C786.203 165.402 787.2 165.365 788.3 165.292C789.4 165.212 790.54 165.108 791.72 164.982C792.893 164.848 794.05 164.678 795.19 164.472C796.33 164.272 797.383 164.022 798.35 163.722V172.122C797.383 172.342 796.33 172.535 795.19 172.702C794.05 172.875 792.873 173.025 791.66 173.152C790.447 173.285 789.243 173.378 788.05 173.432C786.857 173.492 785.737 173.522 784.69 173.522ZM785.25 150.282C783.757 150.282 782.507 150.368 781.5 150.542C780.493 150.708 779.673 150.958 779.04 151.292C778.407 151.625 777.92 152.035 777.58 152.522C777.247 153.008 777.007 153.568 776.86 154.202C777.94 154.502 779.103 154.718 780.35 154.852C781.603 154.978 783.05 155.042 784.69 155.042C786.337 155.042 787.627 154.978 788.56 154.852C789.487 154.718 790.177 154.548 790.63 154.342C791.077 154.135 791.357 153.912 791.47 153.672C791.577 153.432 791.63 153.178 791.63 152.912C791.63 152.652 791.583 152.365 791.49 152.052C791.397 151.732 791.137 151.442 790.71 151.182C790.283 150.922 789.64 150.708 788.78 150.542C787.92 150.368 786.743 150.282 785.25 150.282Z" />
-                <path d="M829.98 143.682V152.132C828.267 151.832 826.53 151.682 824.77 151.682C821.75 151.682 819.01 152.075 816.55 152.862V172.402H807.59V144.412H809.83L812.85 146.202C814.643 145.228 816.547 144.498 818.56 144.012C820.573 143.532 822.813 143.292 825.28 143.292C826.14 143.292 826.96 143.328 827.74 143.402C828.527 143.475 829.273 143.568 829.98 143.682Z" />
-                <path d="M850.92 165.962C852.333 165.962 853.453 165.905 854.28 165.792C855.1 165.678 855.713 165.538 856.12 165.372C856.533 165.205 856.797 165.018 856.91 164.812C857.017 164.605 857.07 164.392 857.07 164.172C857.07 163.312 856.56 162.788 855.54 162.602C854.513 162.415 852.88 162.285 850.64 162.212C849.893 162.212 849.007 162.175 847.98 162.102C846.953 162.022 845.89 161.888 844.79 161.702C843.69 161.515 842.597 161.235 841.51 160.862C840.43 160.488 839.47 159.968 838.63 159.302C837.79 158.628 837.11 157.788 836.59 156.782C836.063 155.775 835.8 154.542 835.8 153.082V152.522C835.8 151.295 836.033 150.128 836.5 149.022C836.967 147.922 837.78 146.942 838.94 146.082C840.093 145.228 841.633 144.548 843.56 144.042C845.48 143.542 847.897 143.292 850.81 143.292H852.37C854.497 143.292 856.57 143.412 858.59 143.652C860.603 143.892 862.56 144.238 864.46 144.692V152.472C862.227 151.945 860.007 151.542 857.8 151.262C855.6 150.982 853.323 150.842 850.97 150.842C849.517 150.842 848.38 150.898 847.56 151.012C846.74 151.125 846.123 151.275 845.71 151.462C845.303 151.648 845.043 151.845 844.93 152.052C844.817 152.252 844.76 152.465 844.76 152.692C844.76 153.478 845.263 153.972 846.27 154.172C847.277 154.378 848.92 154.518 851.2 154.592C851.907 154.632 852.773 154.672 853.8 154.712C854.827 154.745 855.89 154.865 856.99 155.072C858.09 155.278 859.183 155.568 860.27 155.942C861.35 156.315 862.32 156.835 863.18 157.502C864.033 158.175 864.723 159.015 865.25 160.022C865.77 161.035 866.03 162.268 866.03 163.722V164.282C866.03 165.515 865.797 166.682 865.33 167.782C864.863 168.882 864.053 169.862 862.9 170.722C861.74 171.575 860.2 172.255 858.28 172.762C856.36 173.268 853.943 173.522 851.03 173.522H849.46C847.033 173.522 844.683 173.398 842.41 173.152C840.13 172.912 837.89 172.568 835.69 172.122V164.332C838.157 164.858 840.703 165.262 843.33 165.542C845.963 165.822 848.493 165.962 850.92 165.962Z" />
-                <path d="M1082.05 12.06H1078.81V2.88H1074.13V0H1086.75V2.88H1082.05V12.06Z" />
-                <path d="M1096.92 9.6H1095.27L1091.88 4.86V12.06H1088.64V0H1091.88L1096.09 6.18L1100.31 0H1103.55V12.06H1100.31V4.86L1096.92 9.6Z" />
-            </g>
-            <defs>
-                <clipPath id="clip0_1_2">
-                    <rect width="1104" height="174" fill="white" />
-                </clipPath>
-            </defs>
-        </svg>
+
+            <svg
+                id="logo"
+                width="300"
+                height="48"
+                viewBox="0 0 1104 174"
+                fill="none"
+                xmlns="http://www.w3.org/2000/svg"
+            >
+                <g clip-path="url(#clip0_1_2)">
+                    <path d="M20.12 104.429H0.679993V99.6288L36.44 24.0288H58.04L93.8 99.6288V104.429H74.36L67.4 88.8288H26.96L20.12 104.429ZM35 70.8288H59.48L47.24 43.3488L35 70.8288Z" />
+                    <path d="M168.8 24.0288H190.4V70.9488C190.4 75.9088 189.76 80.5488 188.48 84.8688C187.2 89.1888 185 92.9888 181.88 96.2688C178.76 99.5488 174.56 102.129 169.28 104.009C164 105.889 157.36 106.829 149.36 106.829H145.04C137.04 106.829 130.4 105.889 125.12 104.009C119.84 102.129 115.64 99.5488 112.52 96.2688C109.4 92.9888 107.2 89.1888 105.92 84.8688C104.64 80.5488 104 75.9088 104 70.9488V24.0288H125.6V67.8288C125.6 70.4688 125.72 72.9688 125.96 75.3288C126.2 77.6888 126.98 79.7888 128.3 81.6288C129.62 83.4688 131.74 84.9288 134.66 86.0088C137.58 87.0888 141.76 87.6288 147.2 87.6288C152.56 87.6288 156.72 87.0888 159.68 86.0088C162.64 84.9288 164.78 83.4688 166.1 81.6288C167.42 79.7888 168.2 77.6888 168.44 75.3288C168.68 72.9688 168.8 70.4688 168.8 67.8288V24.0288Z" />
+                    <path d="M289.4 104.429H268.52L252.44 79.2288C248.12 79.2288 243.96 79.1888 239.96 79.1088C235.96 79.0288 232.04 78.6688 228.2 78.0288V104.429H206.6V24.0288C213.48 23.2288 220.7 22.6288 228.26 22.2288C235.82 21.8288 243.24 21.6288 250.52 21.6288H253.88C260.6 21.6288 266.16 22.3288 270.56 23.7288C274.96 25.1288 278.48 27.0288 281.12 29.4288C283.76 31.8288 285.6 34.6288 286.64 37.8288C287.68 41.0288 288.2 44.4288 288.2 48.0288V52.8288C288.2 57.8688 287.2 62.4488 285.2 66.5688C283.2 70.6888 279.64 73.8688 274.52 76.1088L289.4 99.6288V104.429ZM250.76 40.8288C247.96 40.8288 244.5 40.9088 240.38 41.0688C236.26 41.2288 232.2 41.5088 228.2 41.9088V58.9488C231.96 59.4288 235.52 59.7288 238.88 59.8488C242.24 59.9688 244.96 60.0288 247.04 60.0288C251.84 60.0288 255.56 59.8088 258.2 59.3688C260.84 58.9288 262.78 58.2688 264.02 57.3888C265.26 56.5088 266 55.3888 266.24 54.0288C266.48 52.6688 266.6 51.0288 266.6 49.1088C266.6 48.0688 266.52 47.0488 266.36 46.0488C266.2 45.0488 265.62 44.1488 264.62 43.3488C263.62 42.5488 262.04 41.9288 259.88 41.4888C257.72 41.0488 254.68 40.8288 250.76 40.8288Z" />
+                    <path d="M368.48 24.0288H390.08V70.9488C390.08 75.9088 389.44 80.5488 388.16 84.8688C386.88 89.1888 384.68 92.9888 381.56 96.2688C378.44 99.5488 374.24 102.129 368.96 104.009C363.68 105.889 357.04 106.829 349.04 106.829H344.72C336.72 106.829 330.08 105.889 324.8 104.009C319.52 102.129 315.32 99.5488 312.2 96.2688C309.08 92.9888 306.88 89.1888 305.6 84.8688C304.32 80.5488 303.68 75.9088 303.68 70.9488V24.0288H325.28V67.8288C325.28 70.4688 325.4 72.9688 325.64 75.3288C325.88 77.6888 326.66 79.7888 327.98 81.6288C329.3 83.4688 331.42 84.9288 334.34 86.0088C337.26 87.0888 341.44 87.6288 346.88 87.6288C352.24 87.6288 356.4 87.0888 359.36 86.0088C362.32 84.9288 364.46 83.4688 365.78 81.6288C367.1 79.7888 367.88 77.6888 368.12 75.3288C368.36 72.9688 368.48 70.4688 368.48 67.8288V24.0288Z" />
+                    <path d="M461.48 87.9888H450.44L427.88 56.4288V104.429H406.28V24.0288H427.88L455.96 65.1888L484.04 24.0288H505.64V104.429H484.04V56.4288L461.48 87.9888Z" />
+                    <path d="M560.96 87.6288C565.28 87.6288 568.66 87.4088 571.1 86.9688C573.54 86.5288 575.38 85.9688 576.62 85.2888C577.86 84.6088 578.64 83.8488 578.96 83.0088C579.28 82.1688 579.44 81.3888 579.44 80.6688C579.44 79.2288 579.1 78.0888 578.42 77.2488C577.74 76.4088 576.66 75.7488 575.18 75.2688C573.7 74.7888 571.74 74.4488 569.3 74.2488C566.86 74.0488 563.92 73.9088 560.48 73.8288C558.48 73.7488 556.12 73.6288 553.4 73.4688C550.68 73.3088 547.88 72.9688 545 72.4488C542.12 71.9288 539.28 71.1688 536.48 70.1688C533.68 69.1688 531.18 67.7688 528.98 65.9688C526.78 64.1688 525 61.9088 523.64 59.1888C522.28 56.4688 521.6 53.1488 521.6 49.2288V47.6688C521.6 44.2288 522.22 40.9288 523.46 37.7688C524.7 34.6088 526.82 31.8288 529.82 29.4288C532.82 27.0288 536.86 25.1288 541.94 23.7288C547.02 22.3288 553.36 21.6288 560.96 21.6288H565.28C570.8 21.6288 576.16 21.9688 581.36 22.6488C586.56 23.3288 591.64 24.3088 596.6 25.5888V45.3888C590.76 44.0288 584.98 42.9488 579.26 42.1488C573.54 41.3488 567.72 40.9488 561.8 40.9488C557.48 40.9488 554.08 41.1688 551.6 41.6088C549.12 42.0488 547.26 42.6088 546.02 43.2888C544.78 43.9688 544 44.6888 543.68 45.4488C543.36 46.2088 543.2 46.9488 543.2 47.6688C543.2 49.1088 543.54 50.2688 544.22 51.1488C544.9 52.0288 546 52.7088 547.52 53.1888C549.04 53.6688 551 54.0088 553.4 54.2088C555.8 54.4088 558.68 54.5488 562.04 54.6288C563.96 54.7088 566.24 54.8088 568.88 54.9288C571.52 55.0488 574.3 55.3688 577.22 55.8888C580.14 56.4088 583.02 57.1888 585.86 58.2288C588.7 59.2688 591.24 60.6888 593.48 62.4888C595.72 64.2888 597.54 66.5688 598.94 69.3288C600.34 72.0888 601.04 75.4288 601.04 79.3488V80.9088C601.04 84.3488 600.42 87.6288 599.18 90.7488C597.94 93.8688 595.8 96.6288 592.76 99.0288C589.72 101.429 585.68 103.329 580.64 104.729C575.6 106.129 569.24 106.829 561.56 106.829H557.24C550.92 106.829 544.78 106.489 538.82 105.809C532.86 105.129 527 104.189 521.24 102.989V82.8288C527.72 84.2688 534.42 85.4288 541.34 86.3088C548.26 87.1888 554.8 87.6288 560.96 87.6288Z" />
+                    <path d="M616.64 24.0288H638.24V55.2288H652.04L678.2 24.0288H699.32V28.8288L669.68 64.2288L699.32 99.6288V104.429H678.2L653 74.4288H638.24V104.429H616.64V24.0288Z" />
+                    <path d="M741.2 104.429V74.5488L709.76 28.8288V24.0288H731L752 57.3888L773.12 24.0288H794.24V28.8288L762.8 74.5488V104.429H741.2Z" />
+                    <path d="M826.64 24.0288L868.76 75.6288V24.0288H890.36V104.429H868.76L826.64 52.8288V104.429H805.04V24.0288H826.64Z" />
+                    <path d="M984.08 24.0288V43.2288H928.76V55.2288H972.08V74.4288H928.76V85.2288H984.08V104.429H907.16V24.0288H984.08Z" />
+                    <path d="M1048.4 104.429H1026.8V43.2288H995.6V24.0288H1079.72V43.2288H1048.4V104.429Z" />
+                    <path d="M214.65 172.402V141.662C214.65 139.875 214.9 138.288 215.4 136.902C215.907 135.522 216.7 134.358 217.78 133.412C218.867 132.458 220.287 131.728 222.04 131.222C223.793 130.722 225.94 130.472 228.48 130.472C229.52 130.472 230.527 130.518 231.5 130.612C232.467 130.705 233.323 130.845 234.07 131.032V139.202C233.323 138.975 232.493 138.815 231.58 138.722C230.667 138.628 229.78 138.582 228.92 138.582C227.653 138.582 226.667 138.675 225.96 138.862C225.247 139.048 224.713 139.328 224.36 139.702C224.007 140.075 223.793 140.515 223.72 141.022C223.647 141.522 223.61 142.108 223.61 142.782V144.412H232.23V152.242H223.61V172.402H214.65Z" />
+                    <path d="M261.5 143.682V152.132C259.787 151.832 258.053 151.682 256.3 151.682C253.273 151.682 250.53 152.075 248.07 152.862V172.402H239.11V144.412H241.35L244.37 146.202C246.163 145.228 248.067 144.498 250.08 144.012C252.1 143.532 254.34 143.292 256.8 143.292C257.66 143.292 258.48 143.328 259.26 143.402C260.047 143.475 260.793 143.568 261.5 143.682Z" />
+                    <path d="M267.1 159.522V157.282C267.1 155.342 267.343 153.522 267.83 151.822C268.317 150.128 269.173 148.645 270.4 147.372C271.633 146.105 273.277 145.108 275.33 144.382C277.383 143.655 279.977 143.292 283.11 143.292H284.68C287.813 143.292 290.407 143.655 292.46 144.382C294.513 145.108 296.157 146.105 297.39 147.372C298.617 148.645 299.473 150.128 299.96 151.822C300.447 153.522 300.69 155.342 300.69 157.282V159.522C300.69 161.462 300.447 163.282 299.96 164.982C299.473 166.682 298.617 168.165 297.39 169.432C296.157 170.698 294.513 171.695 292.46 172.422C290.407 173.155 287.813 173.522 284.68 173.522H283.11C279.977 173.522 277.383 173.155 275.33 172.422C273.277 171.695 271.633 170.698 270.4 169.432C269.173 168.165 268.317 166.682 267.83 164.982C267.343 163.282 267.1 161.462 267.1 159.522ZM276.06 158.402C276.06 159.295 276.097 160.172 276.17 161.032C276.243 161.892 276.533 162.638 277.04 163.272C277.54 163.905 278.313 164.418 279.36 164.812C280.407 165.205 281.92 165.402 283.9 165.402C285.84 165.402 287.34 165.205 288.4 164.812C289.467 164.418 290.25 163.905 290.75 163.272C291.257 162.638 291.547 161.892 291.62 161.032C291.693 160.172 291.73 159.295 291.73 158.402C291.73 157.508 291.693 156.632 291.62 155.772C291.547 154.912 291.257 154.165 290.75 153.532C290.25 152.898 289.467 152.385 288.4 151.992C287.34 151.598 285.84 151.402 283.9 151.402C281.92 151.402 280.407 151.598 279.36 151.992C278.313 152.385 277.54 152.898 277.04 153.532C276.533 154.165 276.243 154.912 276.17 155.772C276.097 156.632 276.06 157.508 276.06 158.402Z" />
+                    <path d="M307.69 144.412H309.93L312.67 146.032C314.497 145.138 316.437 144.458 318.49 143.992C320.543 143.525 322.763 143.292 325.15 143.292C327.73 143.292 329.903 143.572 331.67 144.132C333.443 144.692 334.927 145.455 336.12 146.422C337.987 145.375 339.947 144.592 342 144.072C344.053 143.552 346.313 143.292 348.78 143.292C351.727 143.292 354.18 143.655 356.14 144.382C358.1 145.108 359.647 146.095 360.78 147.342C361.92 148.595 362.723 150.078 363.19 151.792C363.657 153.512 363.89 155.342 363.89 157.282V172.402H354.93V158.402C354.93 157.508 354.893 156.642 354.82 155.802C354.747 154.962 354.513 154.215 354.12 153.562C353.727 152.908 353.103 152.385 352.25 151.992C351.39 151.598 350.177 151.402 348.61 151.402C346.97 151.402 345.43 151.525 343.99 151.772C342.55 152.012 341.16 152.375 339.82 152.862C340.007 153.568 340.127 154.295 340.18 155.042C340.24 155.788 340.27 156.535 340.27 157.282V172.402H331.31V158.402C331.31 157.508 331.273 156.642 331.2 155.802C331.127 154.962 330.893 154.215 330.5 153.562C330.107 152.908 329.48 152.385 328.62 151.992C327.767 151.598 326.553 151.402 324.98 151.402C321.92 151.402 319.14 151.852 316.64 152.752V172.402H307.69V144.412Z" />
+                    <path d="M400.11 165.962C401.53 165.962 402.65 165.905 403.47 165.792C404.29 165.678 404.907 165.538 405.32 165.372C405.727 165.205 405.987 165.018 406.1 164.812C406.213 164.605 406.27 164.392 406.27 164.172C406.27 163.312 405.757 162.788 404.73 162.602C403.703 162.415 402.07 162.285 399.83 162.212C399.083 162.212 398.197 162.175 397.17 162.102C396.143 162.022 395.08 161.888 393.98 161.702C392.88 161.515 391.79 161.235 390.71 160.862C389.623 160.488 388.66 159.968 387.82 159.302C386.98 158.628 386.3 157.788 385.78 156.782C385.26 155.775 385 154.542 385 153.082V152.522C385 151.295 385.233 150.128 385.7 149.022C386.16 147.922 386.97 146.942 388.13 146.082C389.29 145.228 390.83 144.548 392.75 144.042C394.67 143.542 397.087 143.292 400 143.292H401.57C403.697 143.292 405.767 143.412 407.78 143.652C409.793 143.892 411.753 144.238 413.66 144.692V152.472C411.42 151.945 409.2 151.542 407 151.262C404.793 150.982 402.517 150.842 400.17 150.842C398.71 150.842 397.57 150.898 396.75 151.012C395.93 151.125 395.313 151.275 394.9 151.462C394.493 151.648 394.233 151.845 394.12 152.052C394.007 152.252 393.95 152.465 393.95 152.692C393.95 153.478 394.453 153.972 395.46 154.172C396.473 154.378 398.117 154.518 400.39 154.592C401.097 154.632 401.963 154.672 402.99 154.712C404.017 154.745 405.08 154.865 406.18 155.072C407.287 155.278 408.38 155.568 409.46 155.942C410.54 156.315 411.51 156.835 412.37 157.502C413.23 158.175 413.92 159.015 414.44 160.022C414.96 161.035 415.22 162.268 415.22 163.722V164.282C415.22 165.515 414.987 166.682 414.52 167.782C414.06 168.882 413.25 169.862 412.09 170.722C410.93 171.575 409.39 172.255 407.47 172.762C405.55 173.268 403.133 173.522 400.22 173.522H398.65C396.23 173.522 393.88 173.398 391.6 173.152C389.327 172.912 387.087 172.568 384.88 172.122V164.332C387.347 164.858 389.893 165.262 392.52 165.542C395.153 165.822 397.683 165.962 400.11 165.962Z" />
+                    <path d="M438.74 173.522H437.67C434.537 173.522 431.943 173.155 429.89 172.422C427.837 171.695 426.197 170.698 424.97 169.432C423.737 168.165 422.877 166.682 422.39 164.982C421.903 163.282 421.66 161.462 421.66 159.522V157.282C421.66 155.342 421.903 153.522 422.39 151.822C422.877 150.128 423.737 148.645 424.97 147.372C426.197 146.105 427.837 145.108 429.89 144.382C431.943 143.655 434.537 143.292 437.67 143.292H439.24C442.413 143.292 445.007 143.525 447.02 143.992C449.033 144.458 450.6 145.102 451.72 145.922C452.84 146.742 453.607 147.692 454.02 148.772C454.427 149.858 454.63 151.015 454.63 152.242V152.802C454.63 153.922 454.49 154.985 454.21 155.992C453.93 156.998 453.27 157.895 452.23 158.682C451.183 159.468 449.643 160.085 447.61 160.532C445.577 160.978 442.807 161.202 439.3 161.202C437.62 161.202 436.07 161.128 434.65 160.982C433.23 160.828 431.887 160.565 430.62 160.192C430.66 160.978 430.79 161.698 431.01 162.352C431.237 163.005 431.657 163.555 432.27 164.002C432.883 164.448 433.75 164.792 434.87 165.032C435.99 165.278 437.483 165.402 439.35 165.402C440.25 165.402 441.25 165.365 442.35 165.292C443.45 165.212 444.587 165.108 445.76 164.982C446.94 164.848 448.097 164.678 449.23 164.472C450.37 164.272 451.427 164.022 452.4 163.722V172.122C451.427 172.342 450.37 172.535 449.23 172.702C448.097 172.875 446.923 173.025 445.71 173.152C444.497 173.285 443.293 173.378 442.1 173.432C440.9 173.492 439.78 173.522 438.74 173.522ZM439.3 150.282C437.807 150.282 436.557 150.368 435.55 150.542C434.537 150.708 433.713 150.958 433.08 151.292C432.447 151.625 431.963 152.035 431.63 152.522C431.29 153.008 431.047 153.568 430.9 154.202C431.98 154.502 433.147 154.718 434.4 154.852C435.647 154.978 437.093 155.042 438.74 155.042C440.38 155.042 441.667 154.978 442.6 154.852C443.533 154.718 444.223 154.548 444.67 154.342C445.117 154.135 445.397 153.912 445.51 153.672C445.623 153.432 445.68 153.178 445.68 152.912C445.68 152.652 445.633 152.365 445.54 152.052C445.447 151.732 445.183 151.442 444.75 151.182C444.323 150.922 443.68 150.708 442.82 150.542C441.967 150.368 440.793 150.282 439.3 150.282Z" />
+                    <path d="M462.19 130.412H464.43L471.15 134.332V172.402H462.19V130.412Z" />
+                    <path d="M478.99 130.412H481.23L487.94 134.332V172.402H478.99V130.412Z" />
+                    <path d="M512.29 173.522H511.23C508.097 173.522 505.503 173.155 503.45 172.422C501.397 171.695 499.753 170.698 498.52 169.432C497.293 168.165 496.437 166.682 495.95 164.982C495.463 163.282 495.22 161.462 495.22 159.522V157.282C495.22 155.342 495.463 153.522 495.95 151.822C496.437 150.128 497.293 148.645 498.52 147.372C499.753 146.105 501.397 145.108 503.45 144.382C505.503 143.655 508.097 143.292 511.23 143.292H512.8C515.973 143.292 518.567 143.525 520.58 143.992C522.593 144.458 524.16 145.102 525.28 145.922C526.4 146.742 527.167 147.692 527.58 148.772C527.987 149.858 528.19 151.015 528.19 152.242V152.802C528.19 153.922 528.05 154.985 527.77 155.992C527.49 156.998 526.83 157.895 525.79 158.682C524.743 159.468 523.203 160.085 521.17 160.532C519.137 160.978 516.363 161.202 512.85 161.202C511.17 161.202 509.623 161.128 508.21 160.982C506.79 160.828 505.447 160.565 504.18 160.192C504.213 160.978 504.343 161.698 504.57 162.352C504.79 163.005 505.21 163.555 505.83 164.002C506.443 164.448 507.31 164.792 508.43 165.032C509.55 165.278 511.043 165.402 512.91 165.402C513.803 165.402 514.8 165.365 515.9 165.292C517.007 165.212 518.147 165.108 519.32 164.982C520.493 164.848 521.65 164.678 522.79 164.472C523.93 164.272 524.983 164.022 525.95 163.722V172.122C524.983 172.342 523.93 172.535 522.79 172.702C521.65 172.875 520.473 173.025 519.26 173.152C518.047 173.285 516.843 173.378 515.65 173.432C514.457 173.492 513.337 173.522 512.29 173.522ZM512.85 150.282C511.357 150.282 510.107 150.368 509.1 150.542C508.093 150.708 507.273 150.958 506.64 151.292C506.007 151.625 505.52 152.035 505.18 152.522C504.847 153.008 504.607 153.568 504.46 154.202C505.54 154.502 506.707 154.718 507.96 154.852C509.207 154.978 510.65 155.042 512.29 155.042C513.937 155.042 515.227 154.978 516.16 154.852C517.093 154.718 517.783 154.548 518.23 154.342C518.677 154.135 518.957 153.912 519.07 153.672C519.183 153.432 519.24 153.178 519.24 152.912C519.24 152.652 519.193 152.365 519.1 152.052C519.007 151.732 518.743 151.442 518.31 151.182C517.883 150.922 517.24 150.708 516.38 150.542C515.52 150.368 514.343 150.282 512.85 150.282Z" />
+                    <path d="M557.58 143.682V152.132C555.867 151.832 554.133 151.682 552.38 151.682C549.353 151.682 546.61 152.075 544.15 152.862V172.402H535.19V144.412H537.43L540.45 146.202C542.243 145.228 544.147 144.498 546.16 144.012C548.173 143.532 550.413 143.292 552.88 143.292C553.74 143.292 554.56 143.328 555.34 143.402C556.127 143.475 556.873 143.568 557.58 143.682Z" />
+                    <path d="M586.13 137.132V144.412H594.75V152.242H586.13V161.202C586.13 161.835 586.167 162.395 586.24 162.882C586.32 163.368 586.527 163.778 586.86 164.112C587.193 164.445 587.717 164.698 588.43 164.872C589.137 165.038 590.123 165.122 591.39 165.122C592.25 165.122 593.147 165.075 594.08 164.982C595.013 164.888 595.853 164.728 596.6 164.502V172.682C595.853 172.868 594.993 173.008 594.02 173.102C593.053 173.195 592.047 173.242 591 173.242C588.467 173.242 586.32 172.998 584.56 172.512C582.807 172.025 581.39 171.325 580.31 170.412C579.23 169.498 578.437 168.358 577.93 166.992C577.423 165.632 577.17 164.075 577.17 162.322V133.212H579.41L586.13 137.132Z" />
+                    <path d="M602.7 159.522V157.282C602.7 155.342 602.943 153.522 603.43 151.822C603.917 150.128 604.773 148.645 606 147.372C607.233 146.105 608.877 145.108 610.93 144.382C612.983 143.655 615.577 143.292 618.71 143.292H620.28C623.413 143.292 626.007 143.655 628.06 144.382C630.113 145.108 631.757 146.105 632.99 147.372C634.217 148.645 635.073 150.128 635.56 151.822C636.047 153.522 636.29 155.342 636.29 157.282V159.522C636.29 161.462 636.047 163.282 635.56 164.982C635.073 166.682 634.217 168.165 632.99 169.432C631.757 170.698 630.113 171.695 628.06 172.422C626.007 173.155 623.413 173.522 620.28 173.522H618.71C615.577 173.522 612.983 173.155 610.93 172.422C608.877 171.695 607.233 170.698 606 169.432C604.773 168.165 603.917 166.682 603.43 164.982C602.943 163.282 602.7 161.462 602.7 159.522ZM611.66 158.402C611.66 159.295 611.697 160.172 611.77 161.032C611.843 161.892 612.133 162.638 612.64 163.272C613.14 163.905 613.913 164.418 614.96 164.812C616.007 165.205 617.52 165.402 619.5 165.402C621.44 165.402 622.94 165.205 624 164.812C625.067 164.418 625.85 163.905 626.35 163.272C626.857 162.638 627.147 161.892 627.22 161.032C627.293 160.172 627.33 159.295 627.33 158.402C627.33 157.508 627.293 156.632 627.22 155.772C627.147 154.912 626.857 154.165 626.35 153.532C625.85 152.898 625.067 152.385 624 151.992C622.94 151.598 621.44 151.402 619.5 151.402C617.52 151.402 616.007 151.598 614.96 151.992C613.913 152.385 613.14 152.898 612.64 153.532C612.133 154.165 611.843 154.912 611.77 155.772C611.697 156.632 611.66 157.508 611.66 158.402Z" />
+                    <path d="M672.51 165.962C673.93 165.962 675.05 165.905 675.87 165.792C676.69 165.678 677.303 165.538 677.71 165.372C678.123 165.205 678.387 165.018 678.5 164.812C678.613 164.605 678.67 164.392 678.67 164.172C678.67 163.312 678.157 162.788 677.13 162.602C676.103 162.415 674.47 162.285 672.23 162.212C671.483 162.212 670.597 162.175 669.57 162.102C668.543 162.022 667.48 161.888 666.38 161.702C665.28 161.515 664.187 161.235 663.1 160.862C662.02 160.488 661.06 159.968 660.22 159.302C659.38 158.628 658.7 157.788 658.18 156.782C657.653 155.775 657.39 154.542 657.39 153.082V152.522C657.39 151.295 657.623 150.128 658.09 149.022C658.557 147.922 659.37 146.942 660.53 146.082C661.683 145.228 663.223 144.548 665.15 144.042C667.07 143.542 669.487 143.292 672.4 143.292H673.96C676.093 143.292 678.167 143.412 680.18 143.652C682.193 143.892 684.153 144.238 686.06 144.692V152.472C683.82 151.945 681.597 151.542 679.39 151.262C677.19 150.982 674.913 150.842 672.56 150.842C671.107 150.842 669.97 150.898 669.15 151.012C668.33 151.125 667.713 151.275 667.3 151.462C666.893 151.648 666.633 151.845 666.52 152.052C666.407 152.252 666.35 152.465 666.35 152.692C666.35 153.478 666.853 153.972 667.86 154.172C668.867 154.378 670.51 154.518 672.79 154.592C673.497 154.632 674.363 154.672 675.39 154.712C676.417 154.745 677.48 154.865 678.58 155.072C679.68 155.278 680.773 155.568 681.86 155.942C682.94 156.315 683.91 156.835 684.77 157.502C685.63 158.175 686.32 159.015 686.84 160.022C687.36 161.035 687.62 162.268 687.62 163.722V164.282C687.62 165.515 687.387 166.682 686.92 167.782C686.453 168.882 685.643 169.862 684.49 170.722C683.33 171.575 681.79 172.255 679.87 172.762C677.95 173.268 675.533 173.522 672.62 173.522H671.05C668.623 173.522 666.273 173.398 664 173.152C661.72 172.912 659.48 172.568 657.28 172.122V164.332C659.747 164.858 662.293 165.262 664.92 165.542C667.553 165.822 670.083 165.962 672.51 165.962Z" />
+                    <path d="M711.13 173.522H710.07C706.937 173.522 704.343 173.155 702.29 172.422C700.237 171.695 698.593 170.698 697.36 169.432C696.133 168.165 695.277 166.682 694.79 164.982C694.303 163.282 694.06 161.462 694.06 159.522V157.282C694.06 155.342 694.303 153.522 694.79 151.822C695.277 150.128 696.133 148.645 697.36 147.372C698.593 146.105 700.237 145.108 702.29 144.382C704.343 143.655 706.937 143.292 710.07 143.292H711.64C714.813 143.292 717.407 143.525 719.42 143.992C721.433 144.458 723 145.102 724.12 145.922C725.24 146.742 726.007 147.692 726.42 148.772C726.827 149.858 727.03 151.015 727.03 152.242V152.802C727.03 153.922 726.89 154.985 726.61 155.992C726.33 156.998 725.67 157.895 724.63 158.682C723.583 159.468 722.043 160.085 720.01 160.532C717.977 160.978 715.203 161.202 711.69 161.202C710.017 161.202 708.47 161.128 707.05 160.982C705.63 160.828 704.287 160.565 703.02 160.192C703.053 160.978 703.183 161.698 703.41 162.352C703.637 163.005 704.057 163.555 704.67 164.002C705.283 164.448 706.15 164.792 707.27 165.032C708.39 165.278 709.883 165.402 711.75 165.402C712.643 165.402 713.643 165.365 714.75 165.292C715.85 165.212 716.987 165.108 718.16 164.982C719.333 164.848 720.49 164.678 721.63 164.472C722.77 164.272 723.823 164.022 724.79 163.722V172.122C723.823 172.342 722.77 172.535 721.63 172.702C720.49 172.875 719.313 173.025 718.1 173.152C716.893 173.285 715.69 173.378 714.49 173.432C713.297 173.492 712.177 173.522 711.13 173.522ZM711.69 150.282C710.203 150.282 708.953 150.368 707.94 150.542C706.933 150.708 706.113 150.958 705.48 151.292C704.847 151.625 704.363 152.035 704.03 152.522C703.69 153.008 703.447 153.568 703.3 154.202C704.38 154.502 705.547 154.718 706.8 154.852C708.047 154.978 709.49 155.042 711.13 155.042C712.777 155.042 714.067 154.978 715 154.852C715.933 154.718 716.623 154.548 717.07 154.342C717.517 154.135 717.797 153.912 717.91 153.672C718.023 153.432 718.08 153.178 718.08 152.912C718.08 152.652 718.033 152.365 717.94 152.052C717.847 151.732 717.583 151.442 717.15 151.182C716.723 150.922 716.08 150.708 715.22 150.542C714.36 150.368 713.183 150.282 711.69 150.282Z" />
+                    <path d="M734.59 130.412H736.83L743.55 134.332V172.402H734.59V130.412Z" />
+                    <path d="M751.38 130.412H753.62L760.34 134.332V172.402H751.38V130.412Z" />
+                    <path d="M784.69 173.522H783.63C780.497 173.522 777.903 173.155 775.85 172.422C773.797 171.695 772.153 170.698 770.92 169.432C769.687 168.165 768.83 166.682 768.35 164.982C767.863 163.282 767.62 161.462 767.62 159.522V157.282C767.62 155.342 767.863 153.522 768.35 151.822C768.83 150.128 769.687 148.645 770.92 147.372C772.153 146.105 773.797 145.108 775.85 144.382C777.903 143.655 780.497 143.292 783.63 143.292H785.2C788.367 143.292 790.96 143.525 792.98 143.992C794.993 144.458 796.56 145.102 797.68 145.922C798.8 146.742 799.567 147.692 799.98 148.772C800.387 149.858 800.59 151.015 800.59 152.242V152.802C800.59 153.922 800.45 154.985 800.17 155.992C799.89 156.998 799.227 157.895 798.18 158.682C797.14 159.468 795.603 160.085 793.57 160.532C791.53 160.978 788.757 161.202 785.25 161.202C783.57 161.202 782.023 161.128 780.61 160.982C779.19 160.828 777.847 160.565 776.58 160.192C776.613 160.978 776.743 161.698 776.97 162.352C777.19 163.005 777.61 163.555 778.23 164.002C778.843 164.448 779.71 164.792 780.83 165.032C781.95 165.278 783.443 165.402 785.31 165.402C786.203 165.402 787.2 165.365 788.3 165.292C789.4 165.212 790.54 165.108 791.72 164.982C792.893 164.848 794.05 164.678 795.19 164.472C796.33 164.272 797.383 164.022 798.35 163.722V172.122C797.383 172.342 796.33 172.535 795.19 172.702C794.05 172.875 792.873 173.025 791.66 173.152C790.447 173.285 789.243 173.378 788.05 173.432C786.857 173.492 785.737 173.522 784.69 173.522ZM785.25 150.282C783.757 150.282 782.507 150.368 781.5 150.542C780.493 150.708 779.673 150.958 779.04 151.292C778.407 151.625 777.92 152.035 777.58 152.522C777.247 153.008 777.007 153.568 776.86 154.202C777.94 154.502 779.103 154.718 780.35 154.852C781.603 154.978 783.05 155.042 784.69 155.042C786.337 155.042 787.627 154.978 788.56 154.852C789.487 154.718 790.177 154.548 790.63 154.342C791.077 154.135 791.357 153.912 791.47 153.672C791.577 153.432 791.63 153.178 791.63 152.912C791.63 152.652 791.583 152.365 791.49 152.052C791.397 151.732 791.137 151.442 790.71 151.182C790.283 150.922 789.64 150.708 788.78 150.542C787.92 150.368 786.743 150.282 785.25 150.282Z" />
+                    <path d="M829.98 143.682V152.132C828.267 151.832 826.53 151.682 824.77 151.682C821.75 151.682 819.01 152.075 816.55 152.862V172.402H807.59V144.412H809.83L812.85 146.202C814.643 145.228 816.547 144.498 818.56 144.012C820.573 143.532 822.813 143.292 825.28 143.292C826.14 143.292 826.96 143.328 827.74 143.402C828.527 143.475 829.273 143.568 829.98 143.682Z" />
+                    <path d="M850.92 165.962C852.333 165.962 853.453 165.905 854.28 165.792C855.1 165.678 855.713 165.538 856.12 165.372C856.533 165.205 856.797 165.018 856.91 164.812C857.017 164.605 857.07 164.392 857.07 164.172C857.07 163.312 856.56 162.788 855.54 162.602C854.513 162.415 852.88 162.285 850.64 162.212C849.893 162.212 849.007 162.175 847.98 162.102C846.953 162.022 845.89 161.888 844.79 161.702C843.69 161.515 842.597 161.235 841.51 160.862C840.43 160.488 839.47 159.968 838.63 159.302C837.79 158.628 837.11 157.788 836.59 156.782C836.063 155.775 835.8 154.542 835.8 153.082V152.522C835.8 151.295 836.033 150.128 836.5 149.022C836.967 147.922 837.78 146.942 838.94 146.082C840.093 145.228 841.633 144.548 843.56 144.042C845.48 143.542 847.897 143.292 850.81 143.292H852.37C854.497 143.292 856.57 143.412 858.59 143.652C860.603 143.892 862.56 144.238 864.46 144.692V152.472C862.227 151.945 860.007 151.542 857.8 151.262C855.6 150.982 853.323 150.842 850.97 150.842C849.517 150.842 848.38 150.898 847.56 151.012C846.74 151.125 846.123 151.275 845.71 151.462C845.303 151.648 845.043 151.845 844.93 152.052C844.817 152.252 844.76 152.465 844.76 152.692C844.76 153.478 845.263 153.972 846.27 154.172C847.277 154.378 848.92 154.518 851.2 154.592C851.907 154.632 852.773 154.672 853.8 154.712C854.827 154.745 855.89 154.865 856.99 155.072C858.09 155.278 859.183 155.568 860.27 155.942C861.35 156.315 862.32 156.835 863.18 157.502C864.033 158.175 864.723 159.015 865.25 160.022C865.77 161.035 866.03 162.268 866.03 163.722V164.282C866.03 165.515 865.797 166.682 865.33 167.782C864.863 168.882 864.053 169.862 862.9 170.722C861.74 171.575 860.2 172.255 858.28 172.762C856.36 173.268 853.943 173.522 851.03 173.522H849.46C847.033 173.522 844.683 173.398 842.41 173.152C840.13 172.912 837.89 172.568 835.69 172.122V164.332C838.157 164.858 840.703 165.262 843.33 165.542C845.963 165.822 848.493 165.962 850.92 165.962Z" />
+                    <path d="M1082.05 12.06H1078.81V2.88H1074.13V0H1086.75V2.88H1082.05V12.06Z" />
+                    <path d="M1096.92 9.6H1095.27L1091.88 4.86V12.06H1088.64V0H1091.88L1096.09 6.18L1100.31 0H1103.55V12.06H1100.31V4.86L1096.92 9.6Z" />
+                </g>
+                <defs>
+                    <clipPath id="clip0_1_2">
+                        <rect width="1104" height="174" fill="white" />
+                    </clipPath>
+                </defs>
+            </svg>
     );
 };
