--- src/widgets/MassAdvert/overallStats/ui/index.ts
@@ -1,1 +1,0 @@
-export * from "./panel"
