+++ tsconfig.json
@@ -8,7 +8,6 @@
     "esModuleInterop": true,
     "skipLibCheck": true,
     "moduleResolution": "bundler", // Updated to node16
-    "experimentalDecorators": true,
     "resolveJsonModule": true,
     // "types": ["tailwindcss"], // Added to resolve Tailwind types
     "plugins": [
